+++ diagnostics/mixed_layer_depth/mixed_layer_depth.html
@@ -32,16 +32,12 @@ was run. For example, this diagnostic's favorite color is "{EXAMPLE_FAV_COLOR}."
 <TH align=left>{CASENAME}
 <TH align=left>OBS
 <TR>
-<TH align=left>Sea ice!
+<TH align=left>Mixed layer depth
 <!-- 
     Relative links to bitmaps of plots made by the diagnostic, see above. 
 -->
-<TH align=center><A href=../seaice_suite/model/detrendedstd_1979-2014.png>plot</A>
-<TH align=center><A href=../seaice_suite/model/detrendedstd_anomalies1979-2014.png>plot</A>
-<TH align=center><A href=../seaice_suite/model/onemolagcorr_1979-2014.png>plot</A>
-<TH align=center><A href=../seaice_suite/model/onemolagcorr_anomalies_1979-2014.png>plot</A>
-<TH align=center><A href=../seaice_suite/model/oneyrlagcorr_1979-2014.png>plot</A>
-<TH align=center><A href=../seaice_suite/model/oneyrlagcorr_anomalies_1979-2014.png>plot</A>
-<TH align=center><A href=../seaice_suite/model/themean_1979-2014.png>plot</A>
+<TH align=center><A href=../mixed_layer_depth/model/modelmldmean.png>MODEL</A>
+<TH align=center><A href=../mixed_layer_depth/obs/obsmldmean.png>OBS</A>
+
 
 </TABLE>
